CreateLocation creates a new location stored in the database as an auto-generated id. This id is stored in the location object when it is created.
Location object offers the option to get all non-String attributes of a Location object as a string.
Location object now offers for the client to create the object with or without the inclusion of an id value.
Location object no longer requires distance nor collected values.
Location object now contains a getDistance method which does not consider the altitude of the user and a given location.

SavedQuestsActivity creates Location object with no id.